+++ iwaswhere-web/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [0.1.21] - 2016-10-14
+### Changed
+- custom fields chart
+
+## [0.1.20] - 2016-10-13
+### Changed
+- custom field stats
+
 ## [0.1.19] - 2016-10-09
 ### Changed
 - capturing custom fields